Move model to zod style
